Reorganize cache tests
