+++ gcsfs/tests/test_dircache.py
@@ -7,6 +7,11 @@ for writes. The HNS-specific overrides for deletes and moves live in
 :class:`gcsfs._dircache.HnsDirCacheUpdater` and are exercised in
 ``test_extended_hns_gcsfs.py``.
 
+Test placement: keep pure in-memory dircache strategy tests here. Filesystem
+integration and bucket-specific routing tests belong in test_core.py,
+test_extended_hns_gcsfs.py, integration/test_extended_hns.py, or
+test_zonal_file.py according to the bucket type they exercise.
+
 The methods under test are pure in-memory operations on ``self.dircache`` (via
 ``invalidate_cache`` / ``_parent``), so no GCS backend is required.
 """
